ACTUALIZACION

+++ imagen/src/main/java/com/smo/imagen/application/ImagenServiceImp.java
@@ -4,6 +4,7 @@ import com.smo.imagen.application.validator.IValidatorImagen;
 import com.smo.imagen.domain.ImagenModel;
 import com.smo.imagen.infrastructure.ImagenRepository;
 import com.smo.imagen.infrastructure.client.CloudinaryServiceClient;
+import org.json.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
@@ -21,7 +22,7 @@ public class ImagenServiceImp implements ImagenService {
     @Autowired
     private IValidatorImagen iValidatorImagen;
 
-
+    JSONObject jsObject;
     @Autowired
     CloudinaryServiceClient cloudinaryServiceClient;
 
@@ -32,13 +33,15 @@ public class ImagenServiceImp implements ImagenService {
 
     public ImagenModel guardarClienteImagen(String cliImgNum, MultipartFile multipartFile) throws IOException {
         iValidatorImagen.validatorImagen(multipartFile);
-
-        cloudinaryServiceClient.uploadImage(multipartFile);
-
+        String cloudinary = cloudinaryServiceClient.uploadImage(multipartFile).getBody().toString();
+        String imgUrl = cloudinary.toString().split(
+                "data")[1].split("url=")[1].split(",")[0].trim();
+        String imgId = cloudinary.split("data")[1].split(
+                        "public_id=")[1].split(",")[0].trim();
         ImagenModel imagenModel = new ImagenModel();
         imagenModel.setCliImgNum(cliImgNum.trim());
-        imagenModel.setCliImgUrl("HOLA URLss");
-        imagenModel.setCloIdImg("HOLA IDss");
+        imagenModel.setCliImgUrl(imgUrl);
+        imagenModel.setCloIdImg(imgId);
         return imagenRepository.save(imagenModel);
     }
 
@@ -57,9 +60,12 @@ public class ImagenServiceImp implements ImagenService {
                 for (ImagenModel cl : list) {
                     listDelete.add(cl.getCloIdImg());
                 }
-                ;
-                imagenRepository.deleteClienteImagenModelByCliImgNum(cliimgNum);
-                return true;
+                if (cloudinaryServiceClient.deleteImagenList(listDelete).getBody().toString().split(",")[1].trim().equals("status=ACCEPTED")) {
+                    imagenRepository.deleteClienteImagenModelByCliImgNum(cliimgNum);
+                    return true;
+                } else {
+                    return false;
+                }
             } else {
                 return false;
             }
@@ -71,6 +77,7 @@ public class ImagenServiceImp implements ImagenService {
     @Override
     public Long eliminarCliImgUnica(String cloIdImg) {
         try {
+            cloudinaryServiceClient.deleteImagen(cloIdImg);
             return imagenRepository.deleteClienteImagenModelByCloIdImg(cloIdImg);
         } catch (Exception e) {
             return imagenRepository.deleteClienteImagenModelByCloIdImg(cloIdImg);
